FIXME: testing only: a testing question in Software, triggerable via D-Bus
to investigate: when I triggered it locally in product selection,
the web frontend was not picking it up.
TO detect such cases, the front end could set an Asked(displayed) property
and the backend could fail a question that was not asked after a
timeout.
Better than infinitely waiting for a question that the user did not see?